chore: split SingleLookuper and MultiLookuper into their own files
